Created user module and controller
